feat(apis): lecture search
