+++ src/apis/rpc/dto.ts
@@ -1,0 +1,7 @@
+import { LectureItem } from '../lecture/types';
+
+/** GET search Resp */
+export interface GetSearchRespDto {
+  /** 课程列表 */
+  data: LectureItem[];
+}
